Simplify json_encode indent parameter
Removed the possibility to enter a string as indentation, as it raises more
questions than answers. So for now, it's only possible to change the indentation
by passing a number to the `indent` parameter.

Perhaps in the future, we could introduce a fifth parameter like `indent_char`,
with which a character can be passed to specify which character needs to be used
for the indentation.

+++ ext/json/json_encoder.c
@@ -59,9 +59,9 @@ static inline void php_json_pretty_print_char(smart_str *buf, int options, char 
 
 static inline void php_json_pretty_print_indent(smart_str *buf, int options, php_json_encoder *encoder) /* {{{ */
 {
-	int i;
+	size_t i;
 
-	int indent_length = encoder->indent_str ? ZSTR_LEN(encoder->indent_str) : 0;
+	size_t indent_length = encoder->indent_str ? ZSTR_LEN(encoder->indent_str) : 0;
 	if (!indent_length) {
 		return;
 	}
